uinify-back/requirements.txt
flask
flask_cors
sqlalchemy
python-dotenv
pydantic
groq
instructor
